src/requirements.txt
matplotlib==3.5.1
numpy==1.22.3
pandas==1.4.1
Pillow==9.2.0
scikit_learn==1.1.1
scipy==1.7.3
torch==1.12.0
torchvision==0.13.0
